Security hardening of the extraction pipeline and runtime mappings

+++ tests/regression/numeral_edge/numeral_edge.h
@@ -640,8 +640,10 @@ struct NumeralEdge {
   static inline const int64_t z_pow2_30 = INT64_C(1073741824);
   /// 5. Numerals in arithmetic expressions
   static inline const uint64_t add_numerals = (UINT64_C(100) + UINT64_C(200));
-  static inline const int64_t mul_numerals = (INT64_C(10) * INT64_C(20));
-  static inline const int64_t sub_numerals = (INT64_C(100) - INT64_C(1));
+  static inline const int64_t mul_numerals = static_cast<int64_t>(
+      static_cast<uint64_t>(INT64_C(10)) * static_cast<uint64_t>(INT64_C(20)));
+  static inline const int64_t sub_numerals = static_cast<int64_t>(
+      static_cast<uint64_t>(INT64_C(100)) - static_cast<uint64_t>(INT64_C(1)));
   /// 6. Numeral as function argument
   static uint64_t take_nat(uint64_t n);
   static inline const uint64_t test_arg = take_nat(UINT64_C(42));
@@ -662,7 +664,10 @@ struct NumeralEdge {
   static bool is_big(uint64_t n);
   /// 12. Multiple Z values in one function
   static inline const int64_t z_arith =
-      (INT64_C(10) + (INT64_C(3) * INT64_C(7)));
+      static_cast<int64_t>(static_cast<uint64_t>(INT64_C(10)) +
+                           static_cast<uint64_t>(static_cast<int64_t>(
+                               static_cast<uint64_t>(INT64_C(3)) *
+                               static_cast<uint64_t>(INT64_C(7)))));
   /// 13. Negative Z in a pair
   static inline const std::pair<int64_t, int64_t> z_pair =
       std::make_pair(INT64_C(-42), INT64_C(42));
